feat: Update getAllServices route to include toggle field , session update
This commit modifies the getAllServices route in the route.ts file to include a toggle field for each service. The toggle field indicates whether the service is included in the user's portal or not. The toggle field is determined based on the user's portal data and the service IDs. If a service ID is present in the portal data, the toggle field is set to true, otherwise it is set to false. This change improves the functionality of the portal by providing a visual indication of the services that are included.

Note: This commit message follows the established conventions in the repository.