fix: Add handleAddService function to AllServicesPage
The handleAddService function is added to the AllServicesPage component to handle the logic for adding a service to the user's personal portal. This function sends a POST request to the server API to add the selected service to the user's personal portal. If the request is successful, the service is added and the page is reloaded. If there is an error, an error message is logged to the console.

Note: This commit message follows the established conventions in the repository.