refactor: Limit services to top 5 in lastAccess route
The code in route.ts has been modified to limit the services returned to the top 5 based on last access time. This change improves the performance of the lastAccess route by reducing the amount of data retrieved and processed.

Note: This commit message follows the established conventions in the repository.